Move keycloak-user-management reducers into keycloak-user-managment
Move reducers from opensrp-store to keycloak-user-management
