+++ packages/keycloak-user-management/src/components/UserList/utils.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { fetchKeycloakUsers, KeycloakUser, removeKeycloakUsers } from '@opensrp/store';
+import { fetchKeycloakUsers, KeycloakUser, removeKeycloakUsers } from '../../ducks/user';
 import { Dictionary } from '@onaio/utils';
 import { TableActions } from './TableActions';
