+++ packages/keycloak-user-management/src/components/forms/UserForm/index.tsx
@@ -2,10 +2,10 @@ import React, { Dispatch, SetStateAction } from 'react';
 import { Button } from 'antd';
 import { Form, Select, Input } from 'formik-antd';
 import { history } from '@onaio/connected-reducer-registry';
-import { KeycloakUser } from '@opensrp/store';
 import { KeycloakService } from '@opensrp/keycloak-service';
 import { Formik } from 'formik';
 import * as Yup from 'yup';
+import { KeycloakUser } from '../../../ducks/user';
 import { URL_ADMIN } from '../../../constants';
 import { submitForm, fetchRequiredActions, UserAction } from './utils';
 import '../../../index.css';
