+++ packages/keycloak-user-management/src/components/forms/UserForm/utils.tsx
@@ -1,9 +1,9 @@
 import { notification } from 'antd';
 import { history } from '@onaio/connected-reducer-registry';
-import { KeycloakUser } from '@opensrp/store';
 import { Dictionary } from '@onaio/utils';
 import { Dispatch, SetStateAction } from 'react';
 import { KeycloakService } from '@opensrp/keycloak-service';
+import { KeycloakUser } from '../../../ducks/user';
 import {
   KEYCLOAK_URL_USERS,
   URL_ADMIN,
