+++ packages/opensrp-store/src/index.tsx
@@ -1,3 +1,2 @@
 export * from './store/store';
-export * from './store/ducks';
 export * from './store/selectors';
